add rules page for competitions
